+++ src/robot/robot_main.cpp
@@ -1,29 +1,37 @@
 #include <Arduino.h>
+#include "remote.h"
 #include "robot_drive.h"
-#include "wireless.h"
-#include "util.h"
 #include "robot_motion_control.h"
+//#include "wireless.h"
+#include "util.h"
 #include "imu.h"
 #include "EulerAngles.h"
-#include <Wire.h>
+//#include <Wire.h>
 #include "Adafruit_VL6180X.h"
 #include "esp_sender.h"
 #include <iostream>
+//#include "remote.h"
+
 
 std::atomic<float> apriltagx{10};
 std::atomic<float> apriltagy{10};
 std::atomic<int> apriltagid{0};
 std::atomic<int> colorid{0}; // color reading from Apriltag
 
+
 void setup() {
     Serial.begin(115200);
     setupDrive();
     resetState();
     Serial.print("resetted");
-    setupWireless();
+    //setupWireless();
 
     // apriltag integration
     setupComm();
+
+    // joystick
+    ums3.begin();
+    initReceiver();
 }
 
 int test = 0;
@@ -35,7 +43,7 @@ void loop() {
     }
     test = test + 1;
 
-
+    //else
     EVERY_N_MILLIS(20) {
         followTrajectory();
     }
@@ -48,7 +56,7 @@ void loop() {
     // Send and print robot values at 20Hz
     EVERY_N_MILLIS(50) {
         updateOdometry();
-        sendRobotData();
+        //sendRobotData();
 
         //Serial.printf("x: %.2f, y: %.2f, theta: %.2f\n",
         //            robotMessage.x, robotMessage.y, robotMessage.theta);
